refactor, move ingest service packages to internal if not directly related to the controlling cli
