--- cmd/ingest-service/daemon/daemon.go
@@ -1,2 +1,0 @@
-// Package daemon is responsible for running the ingest-service in the background.
-package daemon
