+++ internal/server/ingest/ingest.go
@@ -1,0 +1,2 @@
+// Package ingest is responsible for running the ingest-service in the background.
+package ingest
